Se corrigen detalles en el flujo de Categoria, adicionalmente se implementa la funcionalidad de filtrar products por categoria

+++ EcommerceAPI.Infraestructura/Database/Entidades/ProductoEntity.cs
@@ -8,12 +8,13 @@ namespace EcommerceAPI.Infraestructura.Database.Entidades
     {
         [Key]
         public int id_producto {  get; set; }
-        public float valor { get; set; }
+        public decimal valor { get; set; }
         public string descripcion { get; set; }
         public string imagen { get; set; }
         public int stock { get; set; }
         public int id_categoria { get; set; }
         public int id_estado { get; set; }
         public int categoria_padre { get; set; }
+
     }
 }
